Furion pools and money market parts

+++ contracts/furion-swap/interfaces/IFurionSwapFactory.sol
@@ -3,7 +3,6 @@
 pragma solidity ^0.8.10;
 
 interface IFurionSwapFactory {
-
     function getPair(address _token0, address _token1)
         external
         view
@@ -14,10 +13,13 @@ interface IFurionSwapFactory {
         view
         returns (bool);
 
-    function createPair(address _token0, address _token1) external returns (address _pair);
+    function createPair(address _token0, address _token1)
+        external
+        returns (address _pair);
 
-    function allPairs(uint) external view returns (address _pair);
-    function allPairsLength() external view returns (uint);
+    function allPairs(uint256) external view returns (address _pair);
+
+    function allPairsLength() external view returns (uint256);
 
     function incomeMaker() external view returns (address);
